+++ lib/model/filters/date.dart
@@ -122,9 +122,7 @@ class DateFilter extends CollectionFilter {
   }
 
   @override
-  Widget iconBuilder(BuildContext context, double size, {bool showGenericIcon = true}) {
-    return Icon(AIcons.date, size: size);
-  }
+  Widget? iconBuilder(BuildContext context, double size, {bool showGenericIcon = true}) => Icon(AIcons.date, size: size);
 
   @override
   String get category => type;